# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stoxtech Admin API",
        "version": "1.0.0"
    },
    "paths": {
        "/admin/brands/all": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get brands list",
                "description": "Returns brands list data.",
                "operationId": "brands.all",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/brands": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get brands list",
                "description": "Returns brands list data.",
                "operationId": "brands.index",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Brand"
                ],
                "summary": "Store brand",
                "description": "Returns brand data.",
                "operationId": "brands.store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "title",
                                    "slug",
                                    "logo_url"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "logo_url": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "tables": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/brands/{id}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Show brand",
                "description": "Returns brand data.",
                "operationId": "brands.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Brand"
                ],
                "summary": "Update a brand",
                "operationId": "a2095cdf34924899b60bb49de21eb565",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "logo_url": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "tables": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brand"
                ],
                "summary": "Delete a brand",
                "operationId": "8ee4bf224461c4894c6e30148a6523a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized Error"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Brand"
                ],
                "summary": "Update a brand",
                "operationId": "3335140a0f1ae501a78cbd3506a3d173",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "logo_url": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "tables": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Brand": {
                "title": "Brand",
                "description": "Brand model",
                "required": [
                    "id",
                    "name",
                    "title",
                    "slug",
                    "logo_url",
                    "is_active",
                    "tables",
                    "created_at",
                    "updated_at",
                    "deleted_at"
                ],
                "properties": {
                    "id": {
                        "description": "Brand.",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Brand name"
                    },
                    "title": {
                        "type": "string",
                        "example": "Brand title"
                    },
                    "slug": {
                        "type": "string",
                        "example": "brand-slug"
                    },
                    "logo_url": {
                        "type": "string",
                        "example": "/logo-url"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "tables": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "workers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Worker"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Brand"
                }
            },
            "BrandCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Brand"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                },
                "type": "object"
            },
            "BrandResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Brand"
                    }
                },
                "type": "object"
            },
...
            "Permission": {
                "type": "object"
            },
            "Role": {
                "type": "object"
            },
...
            "Worker": {
                "required": [
                    "id",
                    "username",
                    "first_name",
                    "last_name",
                    "email",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Worker ID",
                        "type": "integer"
                    },
                    "username": {
                        "description": "Worker username",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "First name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email"
                    },
                    "is_active": {
                        "description": "Worker activity flag",
                        "type": "boolean"
                    },
                    "target": {
                        "description": "Target amount for the worker",
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-17 08:44:09"
                    },
                    "updated_at": {
                        "description": "Date and time of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-17 08:44:09"
                    },
                    "deleted_at": {
                        "description": "Date and time of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-17 08:44:09",
                        "nullable": true
                    },
                    "last_login": {
                        "description": "Date and time of soft delete",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-12-17 08:44:09",
                        "nullable": true
                    },
                    "permissions": {
                        "description": "Permission array",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    },
                    "roles": {
                        "description": "Array of roles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            },
...
            "Meta": {
                "properties": {
                    "pagination": {
                        "properties": {
                            "total": {
                                "type": "integer"
                            },
                            "count": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "current_page": {
                                "type": "integer"
                            },
                            "total_pages": {
                                "type": "integer"
                            },
                            "links": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token stateless authentication",
                "scheme": "bearer"
            },
            "sanctum_frontend": {
                "type": "apiKey",
                "in": "cookie",
                "name": "laravel_session",
                "description": "Laravel Sanctum stateful authentication"
            },
            "oauth2_google": {
                "type": "oauth2",
                "description": "Google OAuth2 stateful authentication",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "/admin/auth/login/google?redirect_url=http%3A%2F%2Flocalhost&remember_me=1"
                    }
                }
            }
        }
    }
}